chore: 테마 Directory 정리
