+++ app/src/main/java/org/sopt/and/ui/theme/ColorToken.kt
@@ -1,0 +1,14 @@
+package org.sopt.and.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+object WavveColorDarkTokens {
+    val primary = Color.White
+    val secondary = Gray20
+    val tertiary = Gray30
+    val accent = Blue10
+    val background = Black10
+    val onBackground = Color.White
+    val surface = Black20
+    val surfaceVariant = Black30
+}
